requirements.txt
PyQt6>=6.4.0
PySide6>=6.4.0
mysql-connector-python>=8.0.33 